move class define into one folder
