+++ app/src/main/java/com/huozige/lab/container/offlineform/OfflineFormConfigDetailActivity.java
@@ -9,8 +9,8 @@ import androidx.appcompat.app.AlertDialog;
 import com.huozige.lab.container.BaseActivity;
 import com.huozige.lab.container.R;
 import com.huozige.lab.container.proxy.support.offlinecustomform.helper.OfflineFormFileHelper;
-import com.huozige.lab.container.proxy.support.offlinecustomform.model.OfflineFormDefinition;
-import com.huozige.lab.container.proxy.support.offlinecustomform.model.OfflineFormDefinitionFile;
+import com.huozige.lab.container.offlineform.model.OfflineFormDefinition;
+import com.huozige.lab.container.offlineform.model.OfflineFormDefinitionFile;
 
 import org.json.JSONException;
 import org.json.JSONObject;
