+++ app/src/main/java/com/huozige/lab/container/offlineform/model/formitem/FormItemInput.java
@@ -1,4 +1,4 @@
-package com.huozige.lab.container.proxy.support.offlinecustomform.dto;
+package com.huozige.lab.container.offlineform.model.formitem;
 
 import java.util.List;
 
@@ -10,6 +10,6 @@ public class FormItemInput {
     public String itemType;
     public String value;
 
-    public CheckOptionsInput checkOptions;
+    public FormItemValidationInput checkOptions;
     public List<SelectOptionsInput> selectOptionsList;
 }
