+++ app/src/main/java/com/huozige/lab/container/offlineform/model/formitem/FormItemValidationInput.java
@@ -1,0 +1,10 @@
+package com.huozige.lab.container.offlineform.model.formitem;
+
+public class FormItemValidationInput {
+    public int minLength;
+
+    public int maxLength;
+
+    public String regexPattern;
+
+}
