+++ app/src/main/java/com/huozige/lab/container/offlineform/model/formitem/SelectOptionsInput.java
@@ -1,4 +1,4 @@
-package com.huozige.lab.container.proxy.support.offlinecustomform.dto;
+package com.huozige.lab.container.offlineform.model.formitem;
 
 public class SelectOptionsInput {
     public String value;
