+++ app/src/main/java/com/huozige/lab/container/proxy/support/offlinecustomform/FormAdapter.java
@@ -8,7 +8,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.huozige.lab.container.R;
-import com.huozige.lab.container.proxy.support.offlinecustomform.model.BaseFormItem;
+import com.huozige.lab.container.offlineform.model.formitem.BaseFormItem;
 import com.huozige.lab.container.proxy.support.offlinecustomform.viewholder.BaseViewHolder;
 import com.huozige.lab.container.proxy.support.offlinecustomform.viewholder.SelectViewHolder;
 import com.huozige.lab.container.proxy.support.offlinecustomform.viewholder.TextViewHolder;
