--- app/src/main/java/com/huozige/lab/container/proxy/support/offlinecustomform/dto/CheckOptionsInput.java
@@ -1,10 +1,0 @@
-package com.huozige.lab.container.proxy.support.offlinecustomform.dto;
-
-public class CheckOptionsInput {
-    public int minLength;
-
-    public int maxLength;
-
-    public String regexPattern;
-
-}
